+++ src/pages/api/posts/[id]/index.js
@@ -1,13 +1,17 @@
 import nc from "next-connect";
-import Post from "../../../../models/Posts.js";
+import Posts from "../../../../../models/Posts.js";
 // import { knex } from "../../../../knex/knex.js";
 // interaciton with post based on id
 
 const handler = nc()
   .get(async (req, res) => {
-    const post = await Post.query().findById(parseInt(req.query.id)).first();
+    const { id } = req.query;
 
-    res.status(200).json(post);
+    if (!!id) {
+      const post = await Posts.query().findById(parseInt(id)).first();
+
+      res.status(200).json(post);
+    }
   })
   .put(async (req, res) => {
     console.log(req, res);
